Saving weather data to app database

+++ WeatherApp/app/src/main/java/com/example/weatherapp/model/response/onecall/Current.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Current(
     @SerializedName("clouds")
-        val clouds: Double,
+        val clouds: Int,
     @SerializedName("dt")
         val dt: Int,
     @SerializedName("feels_like")
